src/cloud_functions/data_validator/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.4
